+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/doflinx/DOFLinxService.java
@@ -4,8 +4,11 @@ import de.mephisto.vpin.restclient.PreferenceNames;
 import de.mephisto.vpin.restclient.components.ComponentSummary;
 import de.mephisto.vpin.restclient.components.ComponentType;
 import de.mephisto.vpin.restclient.doflinx.DOFLinxSettings;
+import de.mephisto.vpin.restclient.frontend.EmulatorType;
+import de.mephisto.vpin.restclient.frontend.TableDetails;
 import de.mephisto.vpin.restclient.util.DateUtil;
 import de.mephisto.vpin.restclient.util.SystemCommandExecutor;
+import de.mephisto.vpin.server.frontend.popper.pupgames.PUPGameImporter;
 import de.mephisto.vpin.server.games.BackglassNamingHelper;
 import de.mephisto.vpin.server.games.Game;
 import de.mephisto.vpin.server.games.GameEmulator;
@@ -89,8 +92,29 @@ public class DOFLinxService implements InitializingBean, PreferenceChangedListen
   }
 
   @NonNull
-  public String getGameNameForAltSound(@NonNull String gameName) {
-    return FilenameUtils.getBaseName(gameName).replaceAll("Table_", "");
+  public String getGameNameForAltColor(@NonNull Game game) {
+    if (game.getEmulator().getType().equals(EmulatorType.ZenFX) || game.getEmulator().getType().equals(EmulatorType.PinballM)) {
+      return FilenameUtils.getBaseName(game.getGameName()).replaceAll("Table_", "");
+    }
+
+    if (game.getEmulator().getType().equals(EmulatorType.ZenFX3)) {
+      GameEmulator emulator = game.getEmulator();
+      List<TableDetails> tableDetailList = PUPGameImporter.read(emulator.getType(), emulator.getId());
+      Optional<TableDetails> td = tableDetailList.stream().filter(t -> t.getGameName().equals(game.getGameName())).findFirst();
+      String baseName = game.getGameDisplayName();
+      if (baseName.contains("(")) {
+        baseName = baseName.substring(0, baseName.indexOf("(")).trim();
+      }
+      String name = baseName.replaceAll(" ", "_");
+      if (td.isPresent()) {
+        TableDetails tableDetails = td.get();
+        if (tableDetails.getManufacturer() != null) {
+          name = tableDetails.getManufacturer().toUpperCase().replaceAll(" ", "_") + "_" + name;
+        }
+      }
+      return name;
+    }
+    return game.getRom();
   }
 
   private void startDOFLinx() {