+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/frontend/popper/PinUPConnector.java
@@ -34,6 +34,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Service;
 
 import java.awt.event.KeyEvent;
@@ -1590,6 +1592,11 @@ public class PinUPConnector implements FrontendConnector, InitializingBean {
       return EmulatorType.ZenFX;
     }
 
+    //do not mix up pinball m and pinball mania
+    if (emuName.toLowerCase().contains("mania")) {
+      return EmulatorType.OTHER;
+    }
+
     if (emuName.toLowerCase().contains("pinballm")
         || emuName.toLowerCase().contains("pinball m")
         || (launchScript != null && launchScript.toLowerCase().contains("pinballm"))) {
@@ -2688,5 +2695,23 @@ public class PinUPConnector implements FrontendConnector, InitializingBean {
       LOG.error("Failed to initialize PinUPConnector: {}", e.getMessage(), e);
     }
     LOG.info("{} initialization finished.", this.getClass().getSimpleName());
+  }
+
+  @EventListener(ApplicationReadyEvent.class)
+  public void onApplicationReady() {
+    PopperSettings settings = getSettings();
+    //settings can be null for some t
+    if (settings != null && settings.isAutoStart() && getFrontend().getFrontendType().equals(FrontendType.Popper)) {
+      File popperFolder = systemService.getPinupInstallationFolder();
+      File popperMenu = new File(popperFolder, "PinUpMenu.exe");
+      if (popperMenu.exists()) {
+        SystemCommandExecutor executor = new SystemCommandExecutor(Arrays.asList(popperMenu.getName()));
+        executor.setDir(popperFolder);
+        executor.executeCommandAsync();
+      }
+      else {
+        LOG.error("Failed to auto-start PinUP Popper Menu, file not found");
+      }
+    }
   }
 }