+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/GamesResource.java
@@ -1,5 +1,7 @@
 package de.mephisto.vpin.server.games;
 
+import com.sun.jna.platform.win32.User32;
+import com.sun.jna.platform.win32.WinDef;
 import de.mephisto.vpin.restclient.frontend.EmulatorType;
 import de.mephisto.vpin.restclient.frontend.TableDetails;
 import de.mephisto.vpin.restclient.games.GameScoreValidation;
@@ -125,6 +127,28 @@ public class GamesResource {
         SLOG.initLog(game.getId());
         if (vpxService.play(game, altExe, option)) {
           gameStatusService.setActiveStatus(id);
+          new Thread(()-> {
+            Thread.currentThread().setName("Visual Pinball Player Focus Thread");
+            long timeoutMs = 30000;
+            long start = System.currentTimeMillis();
+
+            while (System.currentTimeMillis() - start < timeoutMs) {
+              WinDef.HWND hwnd = User32.INSTANCE.FindWindow(null, "Visual Pinball Player");
+
+              if (hwnd != null) {
+                try { Thread.sleep(4000); } catch (InterruptedException e) { break; }
+                User32.INSTANCE.ShowWindow(hwnd, 9); // SW_RESTORE
+                User32.INSTANCE.SetForegroundWindow(hwnd);
+
+                try { Thread.sleep(4000); } catch (InterruptedException e) { break; }
+                User32.INSTANCE.ShowWindow(hwnd, 9); // SW_RESTORE
+                User32.INSTANCE.SetForegroundWindow(hwnd);
+                return;
+              }
+
+              try { Thread.sleep(500); } catch (InterruptedException e) { break; }
+            }
+          }).start();
           return true;
         }
       }