+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/dialogs/BaseUploadController.java
@@ -13,6 +13,7 @@ import de.mephisto.vpin.ui.util.ProgressDialog;
 import de.mephisto.vpin.ui.util.ProgressResultModel;
 import de.mephisto.vpin.ui.util.StudioFileChooser;
 import de.mephisto.vpin.ui.util.UploadProgressModel;
+import edu.umd.cs.findbugs.annotations.Nullable;
 import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -210,6 +211,12 @@ public abstract class BaseUploadController implements Initializable, DialogContr
 
   public GameEmulatorRepresentation getSelectedEmulator() {
     return emulator;
+  }
+
+  public void setSelectedEmulator(@Nullable GameEmulatorRepresentation emulator) {
+    if(emulator != null) {
+      this.emulatorCombo.setValue(emulator);
+    }
   }
 
   public int getSelectedEmulatorId() {