+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/panels/PlayButtonController.java
@@ -163,7 +163,7 @@ public class PlayButtonController implements Initializable, ChangeListener<Launc
     if (game != null) {
       UISettings uiSettings = client.getPreferenceService().getJsonPreference(PreferenceNames.UI_SETTINGS, UISettings.class);
       if (uiSettings.isHideVPXStartInfo()) {
-        client.getGameService().playGame(game.getId(), altExe, option);
+        launchGame(game, altExe, option);
         return;
       }
 
@@ -179,9 +179,13 @@ public class PlayButtonController implements Initializable, ChangeListener<Launc
           uiSettings.setHideVPXStartInfo(true);
           client.getPreferenceService().setJsonPreference(uiSettings);
         }
-        client.getGameService().playGame(game.getId(), altExe, option);
+        launchGame(game, altExe, option);
       }
     }
+  }
+
+  private static void launchGame(GameRepresentation game, String altExe, String option) {
+    client.getGameService().playGame(game.getId(), altExe, option);
   }
 
   @Override